% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{adjset}
\alias{adjset}
\alias{areReachable}
\alias{find_nearest_adjset}
\alias{find_adjset}
\title{Identify adjustment set}
\usage{
areReachable(G, x, Z, A)

find_nearest_adjset(Gx, x, Z, A)

find_adjset(G, x, y, name, dmat = matrix(0), Gx = matrix(0))
}
\arguments{
\item{G}{a n-by-n adjacency matrix of a graph. \code{G[x, y] = 1} if there is an edge from node \code{x} to \code{y}, zero otherwise.}

\item{x, y}{(integer)
column position of the cause and effect variable, respectively (zero-based index, from 0 to n-1).}

\item{Z}{(integer vector)
column positions of conditioning variables (zero-based index, from 0 to n-1).}

\item{A}{(logical vector)
indicates a subset of nodes paths must be contained in.
The search along a path ends if visiting a node \code{i} such that \code{A[i] = FALSE}.}

\item{Gx}{the backdoor graph of \code{G} w.r.t. \code{x}.}

\item{name}{(string) name of adjustment set, see details.}

\item{dmat}{(optional) a n-by-n adjacency matrix such that \code{dmat[x, y] = 1} if \code{x} is an ancestor of \code{y} in \code{G}.
Added as an optional argument for speed. If not specified, \code{areDescendants} and \code{areAncestors} are called to
identify the relevant relationships.}
}
\value{
\itemize{
\item \code{areReachable}: a logical vector indicating which nodes in \code{G} can be reached from \code{x}
through active paths that traverses only nodes in \code{A} conditional on \code{Z}.
}

\itemize{
\item \code{find_nearest_adjset}: The subset of Z nearest \code{x} valid for adjustment (assuming Z is a valid adjusment set).
}

\itemize{
\item \code{adjset}: an integer vector with the (zero-indexed) column positions of an adjustment set as given by \code{name}.
}
}
\description{
Identify different sets of nodes in a DAG valid for adjustment with respect to
a cause node X and a effect node Y. Based on Zander and Liśkiewicz (2020) algorithm
for finding (minimal) separators and a modified Algorithm 3.1 in Koller (2009, p. 75)
for finding reachable nodes.
}
\details{
The function \code{areReachable} implements Algorithm 3.1 in Koller (2009, p. 75) and
identifies nodes in \code{G} reachable from \code{x} given \code{Z} via an active trail in \code{G}, with the restriction
that all nodes on the trail are in \code{A}.

The function \code{find_nearest_adjset} is used to identify the \code{o-set} and minimal separators.
It assumes \code{Z} is a valid adjustment set (a set of nodes that includes no descendants of \code{x} and separates \code{x} and \code{y} in the backdoor graph),
and returns the subset of \code{Z} reachable from \code{x} given \code{Z}. That is the subset of \code{Z} closest to \code{x}.
A minimal separator is found by applying \code{find_nearest_adjset} again w.r.t \code{y} and the reduced set.

The function \code{find_adjset} applies these algorithms to identify different adjustment set:
\itemize{
\item "o": The optimal adjustment set (o-set). Given \code{A = Anc(x, y)}, the o-set is the set of nodes in \verb{Z = A\\Desc(x)} that can be reached from \code{y} via active paths in \code{A}.
\item "o-min": The minimal O-set. The subset of the o-set reachable from \code{x}, i.e. a minimal separator of \code{x} and \code{y} in the backdoor graph.
\item "pa-min": The minimal parent set. The subset of the parents of \code{x} reachable from \code{y}, i.e. a minimal separator of \code{x} and \code{y} in the backdoor graph.
}

These adjustment sets are defined only when \code{x} is ancestor of \code{y}.
If that is not the case, the function returns \code{y} indicating that $P(y|do(x)) = P(y)$, such that \code{x} has no causal effect on \code{y}.
Additionally,
\itemize{
\item "pa": the parents of \code{x}, read of \code{G} directly.
\item "pa-if": if \code{y} is a descendant of \code{x}, then the set of parents is returned. Otherwise \code{y} is returned.
}
}
